refactor: split shell control command handling
