+++ clients/apple/scripts/architecture-warning-baseline.txt
@@ -1,6 +1,5 @@
 # Exact Apple architecture-warning debt. Keep entries sorted.
 # large entries include their current line ceiling; every reduction tightens it.
-large|Controllers/Shell/ShellHostControlCommandHandling.swift|1688
 large|GhosttyLiveHost.swift|1258
 large|Models/Shell/ShellSettingsSurfaceModel.swift|1209
 large|Models/Shell/ShellSnapshots.swift|2142
